Begin moving DOM manipulation in DOM updates file
